Saving shipping reference to order
Saving shipping reference to order so that when an order is created with table rate shipping, it will be possible to make a refund. Previously we did not have access to the additional "rate id" that table rate plugin adds to the shipping reference. When a refund attempt was made to the shipping it could not find the corresponding row in Collector.