Subtract shipping cost once it is available
Since `$collector_total` only include the products, when comparing with `$wc_total`, it will be smaller, thus the shipping cost must be subtracted from `$wc_total`. The opposite may also be true if shipping is sent both in the fees.shipping and accounted for in `$wc_total`. This will happen when shipping object is sent, while WC has not yet identified the customer's address.